move `common::util::ordered::serde` to `common::util::row_serde`
