+++ src/common/src/util/mod.rs
@@ -29,11 +29,11 @@ mod future_utils;
 pub mod hash_util;
 pub mod iter_util;
 pub mod memcmp_encoding;
-pub mod ordered;
 pub mod pretty_bytes;
 pub mod prost;
 pub mod resource_util;
 pub mod row_id;
+pub mod row_serde;
 pub mod runtime;
 pub mod scan_range;
 pub mod schema_check;
